+++ src/expressions/Expression.cpp
@@ -1,3 +1,3 @@
 #include "../../include/expressions/Expression.h"
 
-int Expression::evaluate() { return 0; }
+float Expression::evaluate() { return 0; }
